Prevent skipped interfaces in the event of a SocketException.
Moving the try/catch into the loop body ensures every interface is given
a chance.

Also this change makes it possible to alter
NetworkIF.setNetworkInterface to throw SocketException rather than
consuming it. Since that method is technically part of the API (although
it shouldn't be), I didn't mess with it.