Refactor LinuxUsbDevice to fix leaking reference
This change fixes a leaking udev device reference when enumerating
devices. Devices with a devtype that was not "usb_device" were never
freed during the enumeration.

The enumeration process has been refactored (in line with the
implementation in LinuxHWDiskStore) to use the more object-oriented
API for udev that's now available from JNA.