[pom] Skip deleted files as that currently fails
let a full release work then maybe delete manually and replace it all as likely old left over items.  no need to delete at least at the moment for manual site push.  Deletion fails which is why this needs skipped as it is too big of a command for git.