[site] Rework site so it works from local (#1925)
* [pom] Use scm plublish instead of site deploy when releasing

* [pom] Drop wagon plugin as not needed and was not used

* [pom] Change scm publish to include staging instead

* [pom] Fix build directory

* [pom] Adjustments to staging including using just 'oshi' for the maven-site folder

* [pom] Drop content from scm publish as that is the default

* [pom] Skip deleted files as that currently fails

let a full release work then maybe delete manually and replace it all as likely old left over items.  no need to delete at least at the moment for manual site push.  Deletion fails which is why this needs skipped as it is too big of a command for git.

* Add javadocs to top-level menu

Co-authored-by: Daniel Widdis <widdis@gmail.com>