Reduce CpuStat.getSystemCpuLoadTicks memory allocation pressure
Add a new method, FileUtil.readLines, which accepts a number of lines to
read and allocates an ArrayList which contains a number of elements
corresponding to the number of lines to be read. If file is not
readable, return Collections.emptyList() to avoid allocating a list.

Use this new method in CpuStat.getSystemCpuLoadTicks to avoid allocating
a String for each line /proc/stat, instead allocating a single String
and ArrayList with a capacity of one.

This reduces allocation of byte[]'s to back the String for each line by
two orders of magnitude (from a few kB to a few tens of bytes) on each
call.